+++ arrow-incubator-test/build.gradle
@@ -1,4 +1,5 @@
 plugins {
+    id "maven-publish"
     id "base"
     id "org.jetbrains.kotlin.jvm"
     id "org.jetbrains.kotlin.kapt"
@@ -8,19 +9,18 @@ plugins {
 }
 
 apply from: "$SUBPROJECT_CONF"
+apply from: "$PUBLISH_CONF"
 
 dependencies {
+    api "io.arrow-kt:arrow-core-test:$VERSION_NAME", excludeArrow
+
     compile project(":arrow-mtl-data")
     compile project(":arrow-free-data")
     compile project(":arrow-recursion-data")
-    compile "io.arrow-kt:arrow-annotations:$VERSION_NAME", excludeArrow
-    compile "io.arrow-kt:arrow-core-data:$VERSION_NAME", excludeArrow
-    compile "io.arrow-kt:arrow-core:$VERSION_NAME", excludeArrow
-    compile "io.arrow-kt:arrow-optics:$VERSION_NAME", excludeArrow
-    compile "io.arrow-kt:arrow-fx:$VERSION_NAME", excludeArrow
     compile "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$KOTLIN_VERSION"
-    compile "org.jetbrains.kotlinx:kotlinx-coroutines-core:$KOTLINX_COROUTINES_VERSION"
     compile "io.kotlintest:kotlintest-runner-junit5:$KOTLIN_TEST_VERSION", excludeArrow
+
     kapt "io.arrow-kt:arrow-meta:$VERSION_NAME"
+
     testRuntime "org.junit.vintage:junit-vintage-engine:$JUNIT_VINTAGE_VERSION"
 }
