--- arrow-incubator-test/src/main/kotlin/arrow/test/concurrency/SideEffect.kt
@@ -1,7 +1,0 @@
-package arrow.test.concurrency
-
-data class SideEffect(var counter: Int = 0) {
-  fun increment() {
-    counter++
-  }
-}
